- Completed! I actually can't believe I can make this work since even the original example contain bugs and I have spent some time on fixing the original example. With the working solution, I have to find a way to load the KTX file where there is no working solution for zig, so I have to port the original sb7ktx framework which fills with full of magical and confusing stuff.
After gone through all those twist and turns, I
can finally reproduce this example in zig. Does
this code looks clean and idiomatic? I never be, but
this is a huge step on learning OpenGL because the
ktx framework is vital in the superbible book,
while it really helps me learn about the file io
in zig.